+++ templates/testimonial-bottom.php
@@ -1,0 +1,7 @@
+<?php
+global $at_template_args;
+
+extract( $at_template_args );
+?>
+
+<div class="bottom_text"><?php echo $atts['bottom_text']; ?></div>