+++ templates/testimonial-close.php
@@ -1,0 +1,7 @@
+<?php
+global $at_template_args;
+
+extract( $at_template_args );
+?>
+
+</div>