+++ templates/testimonial-extra.php
@@ -1,0 +1,7 @@
+<?php
+global $at_template_args;
+
+extract( $at_template_args );
+?>
+
+<div class="extra"><?php echo $testimonial['testimonial_extra']; ?></div>