+++ templates/testimonial-image.php
@@ -1,0 +1,7 @@
+<?php
+global $at_template_args;
+
+extract( $at_template_args );
+?>
+
+<span class="image"><?php echo $testimonial['testimonial_image']; ?></span>