+++ templates/testimonial-open.php
@@ -35,6 +35,10 @@ if ( ! empty( $post_id ) ) {
 	$class = 'testimonials-widget type-testimonials-widget status-publish hentry ' . $class;
 }
 
+if ( $atts['remove_hentry'] ) {
+	$class = str_replace( ' hentry', '', $class );
+}
+
 $class = apply_filters( 'testimonials_widget_get_testimonial_html_class', $class, $testimonial, $atts, $is_list, $is_first, $widget_number );
 ?>